API

LaTeXTables.print2tex — Function

print2tex(A::AbstractVector{T}, rownames::AbstractVector{String}, colnames::AbstractArray{String})

Print a structural result to a latex table.

Examples

TODO: It seems that jldoctest does not generate result files ex1.tex in Documenter.jl.

julia> μ = [[1 2 3 4; 5 6 7 8], [4 3 2 1; 8 7 6 5]]
2-element Vector{Matrix{Int64}}:
 [1 2 3 4; 5 6 7 8]
 [4 3 2 1; 8 7 6 5]

julia> σ = [ones(Int, 2, 4), ones(Int,2,4)]
2-element Vector{Matrix{Int64}}:
 [1 1 1 1; 1 1 1 1]
 [1 1 1 1; 1 1 1 1]

julia> print2tex(μ, σ, ["A", "B"], ["a", "b"], ["1","2"], ["x", "y"], file = "ex1.tex")
noc = 0
14

julia> tex2png("ex1.tex")

print2tex(μ::AbstractVector{<:AbstractMatrix}, σ::AbstractVector{<:AbstractMatrix}, 
          rownames::AbstractVector{<:AbstractString}, colnames::AbstractVector{<:AbstractString},
          subrownames::AbstractVector{<:AbstractString}, subcolnames::AbstractVector{<:AbstractString}; )

Two-level rows and two-level columns where each cell is in the format of μ (σ) with

rownames as the first level row name and subrownames as the second level row name

colnames as the first level column name and subcolnames as the second level column name.

where a typical example is μ denotes the mean value while σ represents the standard error.

The first level column name can be reduced when passing colnames = [""]

Arguments

colnames_of_rownames: the colnames of the columns containing rownames.

isbf: whether to bold cell, the same size as μ

rank_sup: whether to add superscript for each cell, the same size as μ

other_cols, other_col_names, other_cols_σ: columns on the left of the two-level columns

right_cols, right_col_names: columns on the right of the two-level columns

file: tex file path to write

LaTeXTables.tex2png — Function

tex2png(file::String; background::String)

Render a latex table source file into an image via https://quicklatex.com/js/quicklatex.js

By default, the ouput PNG image is transparent, which may make the table hard to read in some image viewers (e.g., VS Code). As a workaround, you can pass background = "white", but this requires convert to be installed.